requirements.txt
flask
numpy<2
pillow
tensorflow-cpu==2.15.0